Adds some introspective tweaks and raising
You can now tell what happened with a function that Mocksie has mocked.
Calls for functions are recorded and can be evaluated if you wish.

Also adds the ability to make an exception raise on a call.  Planning on
adding more raise-ish stuff in the future, so we'll be able to make it
raise on, say the 3rd call, or raise when a certain parameter matches.

Adds support for using a mocksie mock in 'with' statements too.