run_asserts now verifies args and kwargs
Now when run_asserts is called, it will verify the args and kwargs that
were expected.  You can still use the default [kw]arg '*' to bypass
the check and let anything through.

I also cleaned up the assertion text a little to make it more obvious
who was failing to meet expectations and wrapped quotes around the
function name.